Update focus visuals to use cornerRadius and inner/outer strokes.
